Extract roles to lib file
